+++ game/board.py
@@ -1,9 +1,12 @@
 from game.cell import Cell
 from game.models import Tile
+from game.multipliers import Multipliers
+
 class Board:
     def __init__(self):
         self.grid = [[Cell(1, "", "") for _ in range(15)] for _ in range(15)]
-
+        self.fillWithMultipliers()
+        
     @staticmethod
     def calculateWordValue(word):
         result = 0;
@@ -48,64 +51,7 @@ class Board:
             return True;
 
     def fillWithMultipliers(self):
-        doubleLetter = [(4, 1), 
-                        (12, 1),
-                        (1, 4),
-                        (8, 4),
-                        (15, 4),
-                        (3, 7),
-                        (7, 7),
-                        (9, 7),
-                        (13, 7),
-                        (4, 10),
-                        (12, 10),
-                        (0, 12),
-                        (7, 12),
-                        (14, 12),
-                        (3, 15),
-                        (11, 15)]
-        tripleletter = [(6, 2),
-                        (10, 2),
-                        (2, 6),
-                        (6, 6), 
-                        (10, 6), 
-                        (14, 6), 
-                        (1, 8), 
-                        (5, 8), 
-                        (9, 8), 
-                        (13, 8), 
-                        (2, 10), 
-                        (6, 10), 
-                        (10, 10), 
-                        (14, 10), 
-                        (6, 14), 
-                        (10, 14)]
-        doubleWord = [(1, 1), 
-                      (8, 1), 
-                      (15, 1), 
-                      (2, 2), 
-                      (14, 2), 
-                      (3, 3), 
-                      (13, 3), 
-                      (4, 4), 
-                      (12, 4), 
-                      (7, 7), 
-                      (11, 7), 
-                      (4, 12), 
-                      (12, 12), 
-                      (1, 15), 
-                      (8, 15), 
-                      (15, 15)]
-        tripleWord = [(0, 0), 
-                    (7, 0), 
-                    (14, 0), 
-                    (0, 7), 
-                    (14, 7), 
-                    (0, 14), 
-                    (7, 14), 
-                    (14, 14)]
-        for _ in doubleLetter:
-            print(_)
+        pass;
             
     def addTile(self, x, y, tile=Tile):
         self.grid[x][y].addValue(tile)